src/test/javascript/spec/app/entities/player-match-statistic/player-match-statistic.service.spec.ts
/* tslint:disable max-line-length */
import axios from 'axios';

import * as config from '@/shared/config/config';
import {} from '@/shared/date/filters';
import PlayerMatchStatisticService from '@/entities/player-match-statistic/player-match-statistic.service';
import { PlayerMatchStatistic } from '@/shared/model/player-match-statistic.model';

const mockedAxios: any = axios;
jest.mock('axios', () => ({
  get: jest.fn(),
  post: jest.fn(),
  put: jest.fn(),
  delete: jest.fn()
}));

describe('Service Tests', () => {
  describe('PlayerMatchStatistic Service', () => {
    let service: PlayerMatchStatisticService;
    let elemDefault;
    beforeEach(() => {
      service = new PlayerMatchStatisticService();

      elemDefault = new PlayerMatchStatistic(0, 0, 0, 0, 0, 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', async () => {
        const returnedFromService = Object.assign({}, elemDefault);
        mockedAxios.get.mockReturnValue(Promise.resolve({ data: returnedFromService }));

        return service.find(123).then(res => {
          expect(res).toMatchObject(elemDefault);
        });
      });
      it('should create a PlayerMatchStatistic', async () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);

        mockedAxios.post.mockReturnValue(Promise.resolve({ data: returnedFromService }));
        return service.create({}).then(res => {
          expect(res).toMatchObject(expected);
        });
      });

      it('should update a PlayerMatchStatistic', async () => {
        const returnedFromService = Object.assign(
          {
            description: 1,
            value: 1,
            valueDouble: 1,
            valueLong: 1,
            valueString: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);
        mockedAxios.put.mockReturnValue(Promise.resolve({ data: returnedFromService }));

        return service.update(expected).then(res => {
          expect(res).toMatchObject(expected);
        });
      });
      it('should return a list of PlayerMatchStatistic', async () => {
        const returnedFromService = Object.assign(
          {
            description: 1,
            value: 1,
            valueDouble: 1,
            valueLong: 1,
            valueString: 'BBBBBB'
          },
          elemDefault
        );
        const expected = Object.assign({}, returnedFromService);
        mockedAxios.get.mockReturnValue(Promise.resolve([returnedFromService]));
        return service.retrieve({ sort: {}, page: 0, size: 10 }).then(res => {
          expect(res).toContainEqual(expected);
        });
      });
      it('should delete a PlayerMatchStatistic', async () => {
        mockedAxios.delete.mockReturnValue(Promise.resolve({ ok: true }));
        return service.delete(123).then(res => {
          expect(res.ok).toBeTruthy();
        });
      });
    });
  });
});
